Save the SC:R settings in our own settings file.
