+++ client/settings/action-creators.js
@@ -1,6 +1,5 @@
 import { LOCAL_SETTINGS_SET_BEGIN, SCR_SETTINGS_SET_BEGIN } from '../actions'
 import { LOCAL_SETTINGS_MERGE, SCR_SETTINGS_MERGE } from '../../common/ipc-constants'
-import { fromSbToScr } from './settings-records'
 
 const ipcRenderer = IS_ELECTRON ? require('electron').ipcRenderer : null
 
@@ -30,6 +29,6 @@ export function mergeScrSettings(settings) {
     })
 
     // the ipc-handler will dispatch the right UPDATE event (or SET, if there was an error)
-    ipcRenderer.send(SCR_SETTINGS_MERGE, fromSbToScr(settings))
+    ipcRenderer.send(SCR_SETTINGS_MERGE, settings)
   }
 }
